feature: include COVID-19 information in facilities responses (#84)
Gets COVID-19 info for VA health facilities out of a json file and adds it into the response from the facilities API (both application/json and application/geo+json). 

Includes updates to the documentation that will become available on the dev portal.

Collector Health endpoint checks the response of this new data-source as well as the content of it. This will keep us informed if there are any changes to the json that break our ability to load this data.

https://vasdvp.atlassian.net/browse/SHANK-545